clean up repo

+++ src/placer/globalPlacer/GPSimPL/gridbin.h
@@ -35,7 +35,7 @@ class GridBin {
   int URY() { return top; }
   int Height() { return top - bottom; }
   int Width() { return right - left; }
-  unsigned long int Area() { return (unsigned long int) (top - bottom) * (right - left); }
+  unsigned long int Area() { return (unsigned long int) (top - bottom) * (unsigned long int) (right - left); }
   bool IsAllFixedBlk() { return all_terminal; }
   bool OverFill() { return over_fill; }
   void create_adjacent_bin_list(int grid_cnt_x, int grid_cnt_y);
